registration-system/requirements.txt
pymongo>=4.5.0
qrcode[pil]>=7.4.2
python-dotenv>=1.0.0
pillow>=9.0.0
